move go/autoscaler package into go/controller, rename cmd/controller to cmd/autoscaler
