+++ go/cmd/autoscaler/Dockerfile
@@ -1,0 +1,3 @@
+FROM ubuntu
+ADD autoscaler /
+CMD ["/autoscaler"]
