--- go/cmd/controller/Dockerfile
@@ -1,3 +1,0 @@
-FROM ubuntu
-ADD controller /
-CMD ["/controller"]
